requirements.txt
# OCR and PDF handling
paddleocr==3.0.1
paddlepaddle==2.5.0
opencv-contrib-python==4.10.0.84
PyMuPDF==1.26.0
pdf2image==1.17.0
pytesseract==0.3.13
easyocr==1.7.2

# ONNX
onnxruntime==1.22.0

# Core scientific stack
numpy==1.24.4
pillow==11.2.1
pandas==1.5.3
scikit-learn==1.7.0
scikit-image==0.25.2
matplotlib==3.10.1
tqdm==4.67.1

# Deep Learning
torch==2.1.2+cpu
torchvision==0.16.2+cpu

# Utility & system packages
future==1.0.0
PyYAML==6.0.2
openpyxl==3.1.5

# LangChain & AI integration
langchain==0.2.17
langchain-community==0.2.17
langchain-core==0.2.43
langchain-openai==0.1.25
langchain-text-splitters==0.2.4
langsmith==0.1.147
openai==1.63.2

# Other dependencies for tools used
SQLAlchemy==2.0.41
httpx==0.28.1
aiohttp==3.12.11
aiohappyeyeballs==2.6.1
aiosignal==1.3.2
async-timeout==4.0.3
coloredlogs==15.0.1
prettytable==3.16.0
shapely==2.1.1
fsspec==2025.5.1
ujson==5.10.0
regex==2024.11.6
ftfy==6.3.1
einops==0.8.1
tokenizers==0.19.1
tiktoken==0.9.0
pypdfium2==4.30.1
